# openapi-v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/resources": {
      "get": {
        "operationId": "ResourcesV1Controller_listResources",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public resource list"
          }
        },
        "tags": [
          "v1-resources"
        ]
      },
      "post": {
        "operationId": "ResourcesV1WriteController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "resource_type",
                  "version"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "resource_type": {
                    "type": "string",
                    "enum": [
                      "upload",
                      "external"
                    ]
                  },
                  "resource_kind": {
                    "type": "string"
                  },
                  "version": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string",
                    "description": "Legacy Markdown projection; kept for compatibility."
                  },
                  "content_json": {
                    "type": "object",
                    "description": "Canonical Tiptap/ProseMirror document. Submitted as a JSON string in multipart requests."
                  },
                  "preview_draft_id": {
                    "type": "string"
                  },
                  "schematic_code": {
                    "type": "string"
                  },
                  "external_url": {
                    "type": "string",
                    "format": "uri"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Submitted to the existing resource moderation lifecycle."
          }
        },
        "tags": [
          "v1-resource-uploads"
        ],
        "security": [
          {
            "MindAuthBearer": []
          }
        ]
      }
    },
...
    "/v1/resources/drafts/preview": {
      "post": {
        "operationId": "ResourcesV1WriteController_previewDraft",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "resource_kind"
                ],
                "properties": {
                  "resource_kind": {
                    "type": "string",
                    "enum": [
                      "map",
                      "schematic"
                    ]
                  },
                  "schematic_code": {
                    "type": "string",
                    "maxLength": 29360128
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "A private, short-lived, user-bound parsed resource draft."
          }
        },
        "tags": [
          "v1-resource-uploads"
        ],
        "security": [
          {
            "MindAuthBearer": []
          }
        ]
      }
    },
    "/v1/resources/drafts": {
      "post": {
        "operationId": "ResourcesV1WriteController_createUploadDraft",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "resource_kind",
                  "file"
                ],
                "properties": {
                  "resource_kind": {
                    "type": "string"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creates a durable owner-bound upload draft in quarantine."
          }
        },
        "tags": [
          "v1-resource-uploads"
        ],
        "security": [
          {
            "MindAuthBearer": []
          }
        ]
      }
    },
    "/v1/resources/drafts/{draftId}": {
      "get": {
        "operationId": "ResourcesV1WriteController_getUploadDraft",
        "parameters": [
          {
            "name": "draftId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Owner-only draft metadata; quarantine paths are never returned."
          }
        },
        "tags": [
          "v1-resource-uploads"
        ],
        "security": [
          {
            "MindAuthBearer": []
          }
        ]
      },
      "patch": {
        "operationId": "ResourcesV1WriteController_updateUploadDraft",
        "parameters": [
          {
            "name": "draftId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updates editable resource draft metadata."
          }
        },
        "tags": [
          "v1-resource-uploads"
        ],
        "security": [
          {
            "MindAuthBearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ResourcesV1WriteController_deleteUploadDraft",
        "parameters": [
          {
            "name": "draftId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deletes an owner-bound quarantine draft and its private files."
          }
        },
        "tags": [
          "v1-resource-uploads"
        ],
        "security": [
          {
            "MindAuthBearer": []
          }
        ]
      }
    },
    "/v1/resources/drafts/{draftId}/submit": {
      "post": {
        "operationId": "ResourcesV1WriteController_submitUploadDraft",
        "parameters": [
          {
            "name": "draftId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Submits an owner-bound resource draft to existing forum moderation."
          }
        },
        "tags": [
          "v1-resource-uploads"
        ],
        "security": [
          {
            "MindAuthBearer": []
          }
        ]
      }
    },
    "/v1/resources/drafts/{draftId}/preview": {
      "get": {
        "operationId": "ResourcesV1WriteController_draftPreview",
        "parameters": [
          {
            "name": "draftId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Private preview image for the authenticated submitter."
          }
        },
        "tags": [
          "v1-resource-uploads"
        ],
        "security": [
          {
            "MindAuthBearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "MDTBBS Public Client API",
    "description": "Stable V1 contract for web, official applications and third-party public clients. Use /api/v1/capabilities for feature discovery. External server integrations use the separate /api/external/v1 API key boundary. JSON V1 endpoints return the { data, meta } / { error, meta } envelope unless explicitly documented as a raw file or image response.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api",
      "description": "Same-origin MindFourm API"
    }
  ],
  "components": {
    "securitySchemes": {
      "MindAuthBearer": {
        "scheme": "bearer",
        "bearerFormat": "MindAuth access token",
        "type": "http"
      }
    },
...
  }
}